+++ README.md
@@ -1,8 +1,10 @@
-# Version: 1.9.2
+# Ingestion Script
 
 ![Tests](https://github.com/jyablonski/python_docker/actions/workflows/test.yml/badge.svg) ![ECR Deployment](https://github.com/jyablonski/python_docker/actions/workflows/deploy.yml/badge.svg) [![Coverage Status](https://coveralls.io/repos/github/jyablonski/python_docker/badge.svg?branch=master)](https://coveralls.io/github/jyablonski/python_docker?branch=master) ![Code Style](https://img.shields.io/badge/code%20style-black-000000.svg)
 
-### Python Ingestion Resources for [NBA Project](https://github.com/jyablonski/NBA-Dashboard)
+Version: 1.9.2
+
+![NBA ELT Pipeline Data Flow](https://github.com/jyablonski/python_docker/assets/16946556/0b22bce1-6bb4-4bbb-a2a0-a48a16f2ae58)
 
 1. Links to other Repos providing infrastructure for this Project
     * [Shiny Server](https://github.com/jyablonski/NBA-Dashboard)
